fix windows installer exception
